Adding missing implementaion in map_input_array().

+++ dpcpp/kernels.dp.cpp
@@ -208,11 +208,11 @@ void fill_Q (
 void map_input_array (
 	sycl::nd_item<3> item,
 	#ifdef USE_GLOB_SPACE_XMX_INPUTS
-	float *data_to_be_reduced_global,
-	float *data_to_be_reduced_global_arranged
+	bf16 *data_to_be_reduced_global,
+	bf16 *data_to_be_reduced_global_arranged
 	#else
-	float *data_to_be_reduced,
-	float *data_to_be_reduced_arranged
+	bf16 *data_to_be_reduced,
+	bf16 *data_to_be_reduced_arranged
 	#endif
 	#ifdef DEBUG_XMX_INPUTS_INDEX_MAP
 	,
